Allow non-null-assertion
Typescript isn't always able to detect non nullish checks, and you wan't to treat a value as non-nullish by using `!` instead of `as any` or `as unknown as T`, so I suggest disabling this rule from `@typescript-eslint/recommended`. Use with care however, it is a good idea to always check for non-nullish.